requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.0
transformers==4.37.2
diffusers==0.30.3
peft==0.13.2
accelerate==0.34.2
bitsandbytes==0.43.3

opencv-python==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0
imageio==2.35.1
imageio-ffmpeg==0.6.0

decord==0.6.0
moviepy==2.2.1
scikit-video==1.1.11
pydub==0.25.1

timm==1.0.11
open-clip-torch==2.29.0
basicsr==1.4.2
facexlib==0.3.0
pyiqa==0.1.13
lpips==0.1.4
DISTS-pytorch==0.1
vision-aided-loss==0.1.0

numpy==1.25.0
pandas==2.2.3
scipy==1.14.1
einops==0.8.0

fastapi==0.115.4
uvicorn==0.32.0
gradio==5.5.0
streamlit==1.40.0
requests==2.31.0

omegaconf==2.3.0
pydantic==2.9.2
python-dotenv==1.2.1
tqdm==4.66.5
gdown==4.4.0
huggingface-hub==0.26.2

xformers==0.0.21
safetensors==0.4.5
tokenizers==0.15.2
sentencepiece==0.2.0

wandb==0.22.2
tensorboard==2.18.0